# service_api/openapi.json
{"openapi":"3.1.0","info":{"title":"KOSIS News Fact Verification API","description":"Frozen v31b BGE + reranker + PostgreSQL + KOSIS API verification service","version":"1.0.0"},"paths":{"/healthz":{"get":{"summary":"Healthz","operationId":"healthz_healthz_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Healthz Healthz Get"}}}}}}},"/readyz":{"get":{"summary":"Readyz","operationId":"readyz_readyz_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Readyz Readyz Get"}}}}}}},"/v1/verifications":{"post":{"summary":"Create Verification","operationId":"create_verification_v1_verifications_post","parameters":[{"name":"x-api-key","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-Api-Key"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerificationRequest"}}}},"responses":{"202":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/JobAccepted"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/verifications/{job_id}":{"get":{"summary":"Get Verification","operationId":"get_verification_v1_verifications__job_id__get","parameters":[{"name":"job_id","in":"path","required":true,"schema":{"type":"string","title":"Job Id"}},{"name":"x-api-key","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-Api-Key"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/JobView"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/verifications/{job_id}/result":{"get":{"summary":"Get Result","operationId":"get_result_v1_verifications__job_id__result_get","parameters":[{"name":"job_id","in":"path","required":true,"schema":{"type":"string","title":"Job Id"}},{"name":"x-api-key","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-Api-Key"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/verifications/{job_id}/log":{"get":{"summary":"Get Log","operationId":"get_log_v1_verifications__job_id__log_get","parameters":[{"name":"job_id","in":"path","required":true,"schema":{"type":"string","title":"Job Id"}},{"name":"lines","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Lines"}},{"name":"x-api-key","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-Api-Key"}}],"responses":{"200":{"description":"Successful Response","content":{"text/plain":{"schema":{"type":"string"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"JobAccepted":{"properties":{"job_id":{"type":"string","title":"Job Id"},"status":{"$ref":"#/components/schemas/JobStatus"},"status_url":{"type":"string","title":"Status Url"},"result_url":{"type":"string","title":"Result Url"}},"type":"object","required":["job_id","status","status_url","result_url"],"title":"JobAccepted"},"JobStatus":{"type":"string","enum":["QUEUED","RUNNING","SUCCEEDED","FAILED"],"title":"JobStatus"},"JobView":{"properties":{"job_id":{"type":"string","title":"Job Id"},"status":{"$ref":"#/components/schemas/JobStatus"},"input_stage":{"type":"string","title":"Input Stage"},"progress_step":{"type":"string","title":"Progress Step"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"},"started_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Started At"},"completed_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Completed At"},"error":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error"},"result_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Result Url"}},"type":"object","required":["job_id","status","input_stage","progress_step","created_at","updated_at"],"title":"JobView"},"RawClaim":{"properties":{"claim_id":{"anyOf":[{"type":"string","maxLength":100},{"type":"null"}],"title":"Claim Id"},"article_id":{"anyOf":[{"type":"string","maxLength":100},{"type":"null"}],"title":"Article Id"},"title":{"type":"string","maxLength":500,"minLength":1,"title":"Title"},"date":{"type":"string","maxLength":30,"minLength":4,"title":"Date"},"url":{"type":"string","maxLength":2000,"title":"Url","default":""},"claim_text":{"type":"string","maxLength":5000,"minLength":1,"title":"Claim Text"},"prev_sentence":{"type":"string","maxLength":5000,"title":"Prev Sentence","default":"-"},"next_sentence":{"type":"string","maxLength":5000,"title":"Next Sentence","default":"-"},"article_context":{"type":"string","maxLength":20000,"title":"Article Context","default":""}},"type":"object","required":["title","date","claim_text"],"title":"RawClaim"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"VerificationRequest":{"properties":{"input_stage":{"type":"string","enum":["claims","measurements"],"title":"Input Stage","default":"claims"},"claims":{"items":{"$ref":"#/components/schemas/RawClaim"},"type":"array","title":"Claims"},"measurements":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Measurements"},"client_request_id":{"anyOf":[{"type":"string","maxLength":100},{"type":"null"}],"title":"Client Request Id"}},"type":"object","title":"VerificationRequest"}}}}